requirements.txt
aiogram
sqlalchemy
asyncpg
python-dotenv
pillow
google-generativeai
telethon